create rating component, fix route
